Se modificó el constructor de la clase Hotel para que en las habitaciones de los ultimos dos pisos cuesten más, y se modificó una variable en la clase Habitacion

+++ src/hotelito/Habitacion.java
@@ -69,7 +69,7 @@ public class Habitacion {
         return isAvailable;
     }
 
-    public Habitacion(String Tipo, float Costo, int Correlativo, String Piso, boolean isReserved, boolean isAvailable) {
+    public Habitacion(String Tipo, double Costo, int Correlativo, String Piso, boolean isReserved, boolean isAvailable) {
         this.Tipo = Tipo;
         this.Costo = Costo;
         this.Correlativo = Correlativo;
